back/app/sllm/requirements.txt
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
openai>=1.0.0
PyYAML>=6.0
